xa: fix sign warning

+++ modules/demux/xa.c
@@ -212,8 +212,7 @@ static int Control( demux_t *p_demux, int i_query, va_list args )
     demux_sys_t *p_sys  = p_demux->p_sys;
 
     return demux_vaControlHelper( p_demux->s, HEADER_LENGTH,
-                                  p_sys->i_data_size
-                                  ? HEADER_LENGTH + p_sys->i_data_size : -1,
+        p_sys->i_data_size ? (int64_t)HEADER_LENGTH + p_sys->i_data_size : -1,
                                   p_sys->i_bitrate, p_sys->i_frame_size,
                                   i_query, args );
 }
